requirements.txt
fastapi
uvicorn[standard]
httpx

opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http

opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-httpx